requirements.txt
fastapi==0.98.0
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
tweepy==4.14.0